requirements.txt
Django==1.4.3
Jinja2==2.6
Werkzeug==0.8.3
dj-database-url==0.2.1
httplib2==0.7.7
psycopg2==2.4.6
python-instagram==0.8.0
simplejson==2.6.2
wsgiref==0.1.2